Update tool_conversation.py
One bottleneck in the above code is the register_conv_template() function. This function is used to register a chat template. However, this function is very slow because it has to create a new chat object for each chat template. This can significantly affect performance if a large chat template library is used.

One way to fix this bottleneck is to run register_conv_template() once and then save the chat templates to an array. This will make the chat templates accessible more quickly.

It is also possible to run register_conv_template() once and then save the chat templates in a database. This will make the chat templates accessible from anywhere.

Whichever approach is chosen, something must be done to speed up register_conv_template(). Otherwise, it can significantly impact performance if a large chat template library is used.

Below is a code example that can be used to run register_conv_template() once and then register the chat templates into an array: